Move zlib functions to a separate file
