+++ elastic.go
@@ -2,7 +2,6 @@ package main
 
 import (
 	"bytes"
-	"compress/zlib"
 	"context"
 	"crypto/tls"
 	"crypto/x509"
@@ -137,26 +136,6 @@ func findIndices(es elastic.Client, indexGlob string) ([]string, error) {
 
 	return indices, err
 
-}
-
-func newCompressor(key []byte, w io.Writer) (io.WriteCloser, error) {
-
-	zw := zlib.NewWriter(w)
-	_, err := zlib.NewWriterLevel(zw, zlib.BestCompression)
-
-	if err != nil {
-		log.Error("Unable to set zlib writer level", err)
-	}
-	return zw, nil
-}
-
-func newDecompressor(key []byte, r io.Reader) (io.ReadCloser, error) {
-
-	zr, err := zlib.NewReader(r)
-	if err != nil {
-		log.Error("Unable to create zlib reader", err)
-	}
-	return zr, err
 }
 
 func backupDocuments(sb s3Backend, es elastic.Client, keyPath, indexGlob string, batchsize int) error {
